LibWeb: Move gradient parsing code to a separate file
